requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
huggingface_hub
python-dotenv
SpeechRecognition
PyAudio
beautifulsoup4
requests
tiktoken
audio-recorder-streamlit
streamlit_mic_recorder
langchain_google_genai
pathlib
PyPDF2
chromadb
faiss-cpu